refactor(cockpit): extract rust source gate helper
