+++ crates/tokmd-cockpit/src/gates/rust_source.rs
@@ -1,0 +1,28 @@
+/// Check if a file is a relevant Rust source file for mutation testing.
+/// Excludes test files, fuzz targets, etc.
+#[cfg(feature = "git")]
+pub(super) fn is_relevant_rust_source(path: &str) -> bool {
+    let path_lower = path.to_lowercase();
+
+    // Must be a .rs file
+    if !path_lower.ends_with(".rs") {
+        return false;
+    }
+
+    // Exclude test directories
+    if path_lower.contains("/tests/") || path_lower.starts_with("tests/") {
+        return false;
+    }
+
+    // Exclude test files
+    if path_lower.ends_with("_test.rs") || path_lower.ends_with("_tests.rs") {
+        return false;
+    }
+
+    // Exclude fuzz targets
+    if path_lower.contains("/fuzz/") || path_lower.starts_with("fuzz/") {
+        return false;
+    }
+
+    true
+}
